Float division by zero

+++ check_jbossas7.py
@@ -62,9 +62,9 @@ def performance_data(perf_data, params):
                 warning = warning or 0
                 critical = critical or 0
                 data += ";%s;%s" % (warning, critical)
-                
+
             data += " "
-            
+
     return data
 
 
@@ -72,9 +72,9 @@ def numeric_type(param):
     """
     Checks parameter type
     True for float; int or null data; false otherwise
-    
+
     :param param: input param to check
-    """ 
+    """
     if ((type(param) == float or type(param) == int or param == None)):
         return True
     return False
@@ -83,7 +83,7 @@ def numeric_type(param):
 def check_levels(param, warning, critical, message, ok=[]):
     """
     Checks error level
-    
+
     :param param: input param
     :param warning: watermark for warning
     :param critical: watermark for critical
@@ -122,20 +122,20 @@ def get_digest_auth_json(host, port, uri, user, password, payload):
     """
     HTTP GET with Digest Authentication. Returns JSON result.
     Base URI of http://{host}:{port}/management is used
-    
+
     :param host: JBossAS hostname
     :param port: JBossAS HTTP Management Port
     :param uri: URL fragment
     :param user: management username
     :param password: password
-    :param payload: JSON payload 
+    :param payload: JSON payload
     """
     try:
         url = base_url(host, port) + uri
         res = requests.get(url, params=payload, auth=HTTPDigestAuth(user, password))
         data = res.json()
-        
-        try:    
+
+        try:
             outcome = data['outcome']
             if outcome == "failed":
                 print "CRITICAL - Unexpected value : %s" % data
@@ -153,21 +153,21 @@ def post_digest_auth_json(host, port, uri, user, password, payload):
     """
     HTTP POST with Digest Authentication. Returns JSON result.
     Base URI of http://{host}:{port}/management is used
-    
+
     :param host: JBossAS hostname
     :param port: JBossAS HTTP Management Port
     :param uri: URL fragment
     :param user: management username
     :param password: password
-    :param payload: JSON payload 
+    :param payload: JSON payload
     """
     try:
         url = base_url(host, port) + uri
-        headers = {'content-type': 'application/json'}        
+        headers = {'content-type': 'application/json'}
         res = requests.post(url, data=json.dumps(payload), headers=headers, auth=HTTPDigestAuth(user, password))
         data = res.json()
-        
-        try:    
+
+        try:
             outcome = data['outcome']
             if outcome == "failed":
                 print "CRITICAL - Unexpected value : %s" % data
@@ -184,7 +184,7 @@ def post_digest_auth_json(host, port, uri, user, password, payload):
 def base_url(host, port):
     """
     Provides base URL for HTTP Management API
-    
+
     :param host: JBossAS hostname
     :param port: JBossAS HTTP Management Port
     """
@@ -193,12 +193,12 @@ def base_url(host, port):
 
 
 def main(argv):
-    
+
     global ds_stat_types
     ds_stat_types = ['ActiveCount', 'AvailableCount', 'AverageBlockingTime', 'AverageCreationTime',
                      'CreatedCount', 'DestroyedCount', 'MaxCreationTime', 'MaxUsedCount',
                      'MaxWaitTime', 'TimedOut', 'TotalBlockingTime', 'TotalCreationTime']
-    
+
     p = optparse.OptionParser(conflict_handler="resolve", description="This Nagios plugin checks the health of JBossAS.")
 
     p.add_option('-H', '--host', action='store', type='string', dest='host', default='127.0.0.1', help='The hostname you want to connect to')
@@ -228,7 +228,7 @@ def main(argv):
     datasource_name = options.datasource_name
     ds_stat_type = options.ds_stat_type
     thread_stat_type = options.thread_stat_type
-    
+
     if (options.action == 'server_status'):
         warning = str(options.warning or "")
         critical = str(options.critical or "")
@@ -271,7 +271,7 @@ def main(argv):
 
 def exit_with_general_warning(e):
     """
-    
+
     :param e: exception
     """
     if isinstance(e, SystemExit):
@@ -299,15 +299,15 @@ def check_server_status(host, port, user, passwd, warning, critical, perf_data):
     warning = warning or "reload-required"
     critical = critical or ""
     ok = "running"
-    
+
     try:
         payload = {'operation': 'read-attribute', 'name': 'server-state'}
         res = post_digest_auth_json(host, port, "", user, passwd, payload)
         res = res['result']
-        
+
         message = "Server Status '%s'" % res
         message += performance_data(perf_data, [(res, "server_status", warning, critical)])
-    
+
         return check_levels(res, warning, critical, message, ok)
     except Exception, e:
         return exit_with_general_critical(e)
@@ -317,14 +317,14 @@ def get_memory_usage(host, port, user, passwd, is_heap, memory_value):
     try:
         payload = {'include-runtime': 'true'}
         url = "/core-service/platform-mbean/type/memory"
-        
+
         data = get_digest_auth_json(host, port, url, user, passwd, payload)
-        
+
         if is_heap:
             data = data['heap-memory-usage'][memory_value] / (1024 * 1024)
         else:
             data = data['non-heap-memory-usage'][memory_value] / (1024 * 1024)
-        
+
         return data
     except Exception, e:
         return exit_with_general_critical(e)
@@ -332,15 +332,15 @@ def get_memory_usage(host, port, user, passwd, is_heap, memory_value):
 def check_heap_usage(host, port, user, passwd, warning, critical, perf_data):
     warning = warning or 80
     critical = critical or 90
-    
+
     try:
         used_heap = get_memory_usage(host, port, user, passwd, True, 'used')
         max_heap = get_memory_usage(host, port, user, passwd, True, 'max')
         percent = round((float(used_heap * 100) / max_heap), 2)
-        
+
         message = "Heap Memory Utilization %sMB of %sMB" % (used_heap, max_heap)
         message += performance_data(perf_data, [("%.2f%%" % percent, "heap_usage", warning, critical)])
-    
+
         return check_levels(percent, warning, critical, message)
     except Exception, e:
         return exit_with_general_critical(e)
@@ -348,15 +348,15 @@ def check_heap_usage(host, port, user, passwd, warning, critical, perf_data):
 def check_non_heap_usage(host, port, user, passwd, warning, critical, perf_data):
     warning = warning or 80
     critical = critical or 90
-    
+
     try:
         used_heap = get_memory_usage(host, port, user, passwd, False, 'used')
         max_heap = get_memory_usage(host, port, user, passwd, False, 'max')
         percent = round((float(used_heap * 100) / max_heap), 2)
-        
+
         message = "Non Heap Memory Utilization %sMB of %sMB" % (used_heap, max_heap)
         message += performance_data(perf_data, [("%.2f%%" % percent, "non_heap_usage", warning, critical)])
-    
+
         return check_levels(percent, warning, critical, message)
     except Exception, e:
         return exit_with_general_critical(e)
@@ -365,10 +365,10 @@ def get_memory_pool_usage(host, port, user, passwd, pool_name, memory_value):
     try:
         payload = {'include-runtime': 'true', 'recursive':'true'}
         url = "/core-service/platform-mbean/type/memory-pool"
-        
+
         data = get_digest_auth_json(host, port, url, user, passwd, payload)
-        usage = data['name'][pool_name]['usage'][memory_value] / (1024 * 1024)
-        
+        usage = data['name'][pool_name]['usage'][memory_value] / float(1024 * 1024)
+
         return usage
     except Exception, e:
         return exit_with_general_critical(e)
@@ -377,15 +377,15 @@ def get_memory_pool_usage(host, port, user, passwd, pool_name, memory_value):
 def check_eden_space_usage(host, port, user, passwd, memory_pool, warning, critical, perf_data):
     warning = warning or 80
     critical = critical or 90
-    
+
     try:
         used_heap = get_memory_pool_usage(host, port, user, passwd, memory_pool, 'used')
         max_heap = get_memory_pool_usage(host, port, user, passwd, memory_pool, 'max')
         percent = round((float(used_heap * 100) / max_heap), 2)
-        
+
         message = "Eden_Space Utilization %sMB of %sMB" % (used_heap, max_heap)
         message += performance_data(perf_data, [("%.2f%%" % percent, "eden_space_usage", warning, critical)])
-    
+
         return check_levels(percent, warning, critical, message)
     except Exception, e:
         return exit_with_general_critical(e)
@@ -393,15 +393,15 @@ def check_eden_space_usage(host, port, user, passwd, memory_pool, warning, criti
 def check_tenured_gen_usage(host, port, user, passwd, memory_pool, warning, critical, perf_data):
     warning = warning or 80
     critical = critical or 90
-    
+
     try:
         used_heap = get_memory_pool_usage(host, port, user, passwd, memory_pool, 'used')
         max_heap = get_memory_pool_usage(host, port, user, passwd, memory_pool, 'max')
         percent = round((float(used_heap * 100) / max_heap), 2)
-        
+
         message = "Tenured_Gen Utilization %sMB of %sMB" % (used_heap, max_heap)
         message += performance_data(perf_data, [("%.2f%%" % percent, "tenured_gen_usage", warning, critical)])
-    
+
         return check_levels(percent, warning, critical, message)
     except Exception, e:
         return exit_with_general_critical(e)
@@ -409,15 +409,15 @@ def check_tenured_gen_usage(host, port, user, passwd, memory_pool, warning, crit
 def check_survivor_space_usage(host, port, user, passwd, memory_pool, warning, critical, perf_data):
     warning = warning or 80
     critical = critical or 90
-    
+
     try:
         used_heap = get_memory_pool_usage(host, port, user, passwd, memory_pool, 'used')
         max_heap = get_memory_pool_usage(host, port, user, passwd, memory_pool, 'max')
         percent = round((float(used_heap * 100) / max_heap), 2)
-        
+
         message = "Survivor_Space Utilization %sMB of %sMB" % (used_heap, max_heap)
         message += performance_data(perf_data, [("%.2f%%" % percent, "survivor_space_usage", warning, critical)])
-    
+
         return check_levels(percent, warning, critical, message)
     except Exception, e:
         return exit_with_general_critical(e)
@@ -425,15 +425,15 @@ def check_survivor_space_usage(host, port, user, passwd, memory_pool, warning, c
 def check_perm_gen_usage(host, port, user, passwd, memory_pool, warning, critical, perf_data):
     warning = warning or 90
     critical = critical or 95
-    
+
     try:
         used_heap = get_memory_pool_usage(host, port, user, passwd, memory_pool, 'used')
         max_heap = get_memory_pool_usage(host, port, user, passwd, memory_pool, 'max')
         percent = round((float(used_heap * 100) / max_heap), 2)
-        
+
         message = "Perm_Gen Utilization %sMB of %sMB" % (used_heap, max_heap)
         message += performance_data(perf_data, [("%.2f%%" % percent, "perm_gen_usage", warning, critical)])
-    
+
         return check_levels(percent, warning, critical, message)
     except Exception, e:
         return exit_with_general_critical(e)
@@ -441,65 +441,65 @@ def check_perm_gen_usage(host, port, user, passwd, memory_pool, warning, critica
 def check_code_cache_usage(host, port, user, passwd, memory_pool, warning, critical, perf_data):
     warning = warning or 90
     critical = critical or 95
-    
+
     try:
     	if memory_pool == None:
     		memory_pool = 'Code_Cache'
-    		
+
         used_heap = get_memory_pool_usage(host, port, user, passwd, memory_pool, 'used')
         max_heap = get_memory_pool_usage(host, port, user, passwd, memory_pool, 'max')
         percent = round((float(used_heap * 100) / max_heap), 2)
-        
+
         message = "Code_Cache Utilization %sMB of %sMB" % (used_heap, max_heap)
         message += performance_data(perf_data, [("%.2f%%" % percent, "code_cache_usage", warning, critical)])
-    
+
         return check_levels(percent, warning, critical, message)
     except Exception, e:
         return exit_with_general_critical(e)
 
 
 def check_gctime(host, port, user, passwd, memory_pool, warning, critical, perf_data):
-    # Make sure you configure right values for your application    
+    # Make sure you configure right values for your application
     warning = warning or 500
     critical = critical or 1000
-    
+
     try:
         payload = {'include-runtime': 'true', 'recursive':'true'}
         url = "/core-service/platform-mbean/type/garbage-collector"
         res = get_digest_auth_json(host, port, url, user, passwd, payload)
         gc_time = res['name'][memory_pool]['collection-time']
         gc_count = res['name'][memory_pool]['collection-count']
-        
+
         avg_gc_time = 0
-         
+
         if gc_count > 0:
             avg_gc_time = float(gc_time / gc_count)
-        
+
         message = "GC '%s' total-time=%dms count=%s avg-time=%.2fms" % (memory_pool, gc_time, gc_count, avg_gc_time)
         message += performance_data(perf_data, [("%.2fms" % avg_gc_time, "gctime", warning, critical)])
-    
+
         return check_levels(avg_gc_time, warning, critical, message)
     except Exception, e:
         return exit_with_general_critical(e)
-    
+
 
 def check_threading(host, port, user, passwd, thread_stat_type, warning, critical, perf_data):
     warning = warning or 100
     critical = critical or 200
-    
+
     try:
         if thread_stat_type not in ['thread-count', 'peak-thread-count', 'total-started-thread-count', 'daemon-thread-count']:
             return exit_with_general_critical("The thread statistics value type of '%s' is not valid" % thread_stat_type)
-            
+
         payload = {'include-runtime': 'true'}
         url = "/core-service/platform-mbean/type/threading"
-        
+
         data = get_digest_auth_json(host, port, url, user, passwd, payload)
         data = data[thread_stat_type]
-        
+
         message = "Threading Statistics '%s':%s " % (thread_stat_type, data)
         message += performance_data(perf_data, [(data, "threading", warning, critical)])
-    
+
         return check_levels(data, warning, critical, message)
     except Exception, e:
         return exit_with_general_critical(e)
@@ -508,40 +508,40 @@ def check_threading(host, port, user, passwd, thread_stat_type, warning, critica
 def check_queue_depth(host, port, user, passwd, queue_name, warning, critical, perf_data):
     warning = warning or 100
     critical = critical or 200
-    
-    try:    
+
+    try:
         if queue_name is None:
             return exit_with_general_critical("The queue name '%s' is not valid" % queue_name)
-            
+
         payload = {'include-runtime': 'true', 'recursive':'true'}
         url = "/subsystem/messaging/hornetq-server/default/jms-queue/" + queue_name
-        
+
         data = get_digest_auth_json(host, port, url, user, passwd, payload)
         queue_depth = data['message-count']
-        
+
         message = "Queue %s depth %s" % (queue_name, queue_depth)
         message += performance_data(perf_data, [(queue_depth, "queue_depth", warning, critical)])
-    
+
         return check_levels(queue_depth, warning, critical, message)
     except Exception, e:
         return exit_with_general_critical(e)
 
 def get_datasource_stats(host, port, user, passwd, is_xa, ds_name, ds_stat_type):
-    try:    
+    try:
         if ds_name is None:
             return exit_with_general_critical("The ds_name name '%s' is not valid" % ds_name)
         if ds_stat_type not in ds_stat_types:
             return exit_with_general_critical("The datasource statistics type of '%s' is not valid" % ds_stat_type)
-            
+
         payload = {'include-runtime': 'true', 'recursive':'true'}
         if is_xa:
             url = "/subsystem/datasources/xa-data-source/" + ds_name + "/statistics/pool/"
         else:
             url = "/subsystem/datasources/data-source/" + ds_name + "/statistics/pool/"
-        
+
         data = get_digest_auth_json(host, port, url, user, passwd, payload)
         data = data[ds_stat_type]
-        
+
         return data
     except Exception, e:
         return exit_with_general_critical(e)
@@ -550,10 +550,10 @@ def get_datasource_stats(host, port, user, passwd, is_xa, ds_name, ds_stat_type)
 def check_non_xa_datasource(host, port, user, passwd, ds_name, ds_stat_type, warning, critical, perf_data):
     warning = warning or 0
     critical = critical or 10
-    
-    try:    
+
+    try:
         data = get_datasource_stats(host, port, user, passwd, False, ds_name, ds_stat_type)
-        
+
         message = "DataSource %s %s" % (ds_stat_type, data)
         message += performance_data(perf_data, [(data, "datasource", warning, critical)])
         return check_levels(data, warning, critical, message)
@@ -563,8 +563,8 @@ def check_non_xa_datasource(host, port, user, passwd, ds_name, ds_stat_type, war
 def check_xa_datasource(host, port, user, passwd, ds_name, ds_stat_type, warning, critical, perf_data):
     warning = warning or 0
     critical = critical or 10
-    
-    try:    
+
+    try:
         data = get_datasource_stats(host, port, user, passwd, True, ds_name, ds_stat_type)
 
         message = "XA DataSource %s %s" % (ds_stat_type, data)
